requirements.txt
asgiref==3.5.2
coverage==6.4.4
Django==4.1.1
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==14.2.0
phonenumbers==8.12.55
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
